separation of AEC & parallel + linearized reward
